Fix for backwards compatibility of the database: 'zerofill' numeric fields are now parsed just like the v1.4.2+ non-zerofill-ed numeric fields.
For this purpose, the rm0lead() function has been introduced (which is also handy when you need to 'print' a possibly-zero-valued integer, as this function will then spit out "0" instead of the empty string "").
Another bit which was introduced to support this fix is the SQLVALUE_ENUMERATE type, because comment ratings are an 'enumerate' of '1'..'5' and MySQL already showed during previous tests that it did not tolerate
  WHERE field=3
integer-like SQL queries, but required
  WHERE field='3'
instead for this enumerate. Meanwhile we removed the quotes around the 'true' integers to make sure MySQL performs the expected conversions when comparing zerofill-ed numeric fields with literal specified values.