security review: MySQL access + fixed MySQL access class to support arbitrary ASC/DESC sort order combinations.
The previous implementation only allowed for all ASCending or all DESCending sort order; the new implementation looks at the field prefix of the fields listed in the sort order array:a '-' minus means DESCending sort order for that field is required, while no prefix or a '+' plus sign indicates ASCending sort order for that particular field.

All queries have been inspected and action has been taken to ensure:

1) all variable arguments are fed through the SQLValue() filter/converter method.

2) all queries use the Select*() methods whenever possible; under the hood these create the required SELECT FROM queries; only the table joining SELECT FROM ... queries remain as is in the code, invoking the basic Query() method.

All table arguments for the query methods use single quotes for the table name - which makes for easier grep-ping of the code: only one grep to get them all. (Note that the only exception to this rule is those table-joining raw SELECT queries in there!)